Backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
python-dotenv==1.0.1
google-generativeai==0.7.2
psycopg2-binary==2.9.9
